# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AttributionInput": {
        "additionalProperties": false,
        "properties": {
          "clickIdKind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClickIdKind"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "clickIdValue": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Clickidvalue"
          },
          "landingUrl": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Landingurl"
          },
          "referrerUrl": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Referrerurl"
          },
          "utmCampaign": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Utmcampaign"
          },
          "utmContent": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Utmcontent"
          },
          "utmMedium": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Utmmedium"
          },
          "utmSource": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Utmsource"
          },
          "utmTerm": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Utmterm"
          }
        },
        "title": "AttributionInput",
        "type": "object"
      },
      "BudgetBand": {
        "enum": [
          "300000-400000"
        ],
        "title": "BudgetBand",
        "type": "string"
      },
      "ClickIdKind": {
        "enum": [
          "fbclid",
          "ttclid",
          "twclid"
        ],
        "title": "ClickIdKind",
        "type": "string"
      },
      "ConsentInput": {
        "additionalProperties": false,
        "properties": {
          "adMeasurement": {
            "default": false,
            "title": "Admeasurement",
            "type": "boolean"
          },
          "locale": {
            "default": "en-IN",
            "maxLength": 20,
            "minLength": 2,
            "pattern": "^[A-Za-z]{2,3}(?:-[A-Za-z0-9]{2,8})*$",
            "title": "Locale",
            "type": "string"
          },
          "marketing": {
            "default": false,
            "title": "Marketing",
            "type": "boolean"
          },
          "privacyNoticeVersion": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Privacynoticeversion",
            "type": "string"
          },
          "requestedContact": {
            "title": "Requestedcontact",
            "type": "boolean"
          }
        },
        "required": [
          "privacyNoticeVersion",
          "requestedContact"
        ],
        "title": "ConsentInput",
        "type": "object"
      },
...
      "ContactInput": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Email"
          },
          "fullName": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Fullname",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "pattern": "^\\+[1-9]\\d{7,14}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Phone"
          },
          "preferredContactMethod": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreferredContactMethod"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "fullName"
        ],
        "title": "ContactInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InquiryInput": {
        "additionalProperties": false,
        "properties": {
          "budgetBand": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BudgetBand"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InquiryIntent"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "message": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Message"
          },
          "timeframe": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Timeframe"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "title": "InquiryInput",
        "type": "object"
      },
      "InquiryIntent": {
        "enum": [
          "buy",
          "rent",
          "sell",
          "information"
        ],
        "title": "InquiryIntent",
        "type": "string"
      },
...
      "LeadAccepted": {
        "additionalProperties": false,
        "properties": {
          "inquiryId": {
            "format": "uuid",
            "title": "Inquiryid",
            "type": "string"
          },
          "reference": {
            "pattern": "^RE-[A-Z0-9]{10}$",
            "title": "Reference",
            "type": "string"
          },
          "status": {
            "const": "new",
            "title": "Status",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "title": "Submittedat",
            "type": "string"
          }
        },
        "required": [
          "inquiryId",
          "reference",
          "status",
          "submittedAt"
        ],
        "title": "LeadAccepted",
        "type": "object"
      },
      "LeadRequest": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "$ref": "#/components/schemas/AttributionInput"
          },
          "consent": {
            "$ref": "#/components/schemas/ConsentInput"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactInput"
          },
          "formVersion": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Formversion",
            "type": "string"
          },
          "inquiry": {
            "$ref": "#/components/schemas/InquiryInput"
          },
          "propertySlug": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z0-9](?:[a-z0-9-]{0,78}[a-z0-9])?$",
            "title": "Propertyslug",
            "type": "string"
          },
          "website": {
            "default": "",
            "maxLength": 200,
            "title": "Website",
            "type": "string"
          }
        },
        "required": [
          "propertySlug",
          "formVersion",
          "contact",
          "consent"
        ],
        "title": "LeadRequest",
        "type": "object"
      },
...
      "PreferredContactMethod": {
        "enum": [
          "email",
          "phone",
          "whatsapp",
          "no_preference"
        ],
        "title": "PreferredContactMethod",
        "type": "string"
      },
      "PropertyResponse": {
        "additionalProperties": false,
        "properties": {
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "priceLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pricelabel"
          },
          "projectRegistrationNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectregistrationnumber"
          },
          "registrationAuthorityUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registrationauthorityurl"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "summary",
          "locality",
          "priceLabel",
          "projectRegistrationNumber",
          "registrationAuthorityUrl"
        ],
        "title": "PropertyResponse",
        "type": "object"
      },
...
      "Timeframe": {
        "enum": [
          "1-3-months"
        ],
        "title": "Timeframe",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Real-Estate Lead Capture CRM",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/leads": {
      "post": {
        "operationId": "createLead",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadAccepted"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create Lead",
        "tags": [
          "public"
        ]
      }
    },
    "/api/v1/properties/{slug}": {
      "get": {
        "operationId": "getProperty",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "Cache-Control": {
                "description": "Sensitive API responses are private and not stored",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Opaque request correlation identifier",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Property",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}